Refactor SyncServiceLevels
1. Move methods related to SLOConfig object to the
   `controllers/plan/sloConfig.go` to be consistent with the object that
interacts with those methods
2. Clean things up so we are no longer applying non-tagged
   serviceLevelObjectives to the clusters (but still delete ones that
are there)
3. Remove tests related to the un-tagged service level objectives
