+++ controllers/syncer.go
@@ -319,57 +319,16 @@ func (r *ResourceSyncer) delServiceLevels(ctx context.Context) error {
 }
 
 func (r *ResourceSyncer) syncServiceLevels(ctx context.Context) error {
-	if r.picchuConfig.ServiceLevelsFleet != "" && r.picchuConfig.ServiceLevelsNamespace != "" {
-		slos, sloLabels := r.prepareServiceLevelObjectives()
-		if len(slos) > 0 {
-			if err := r.applyDeliveryPlan(ctx, "Ensure Service Levels Namespace", &rmplan.EnsureNamespace{
-				Name: r.picchuConfig.ServiceLevelsNamespace,
-			}); err != nil {
-				return err
-			}
-
-			labels := r.defaultLabels()
-			labels[picchuv1alpha1.LabelTarget] = r.instance.Spec.Target
-
-			if err := r.applyDeliveryPlan(ctx, "Sync App ServiceLevels", &rmplan.SyncServiceLevels{
-				App:                         r.instance.Spec.App,
-				Target:                      r.instance.Spec.Target,
-				Namespace:                   r.picchuConfig.ServiceLevelsNamespace,
-				Labels:                      labels,
-				ServiceLevelObjectiveLabels: sloLabels,
-				ServiceLevelObjectives:      slos,
-			}); err != nil {
-				return err
-			}
-		} else {
 			return r.delServiceLevels(ctx)
 		}
-	} else {
-		r.log.Info("service-levels-fleet and service-levels-namespace not set, skipping SyncServiceLevels")
-	}
-	return nil
-}
 
 func (r *ResourceSyncer) syncSLORules(ctx context.Context) error {
-	slos, labels := r.prepareServiceLevelObjectives()
-	if len(slos) > 0 {
-		if err := r.applyPlan(ctx, "Sync App SLO Rules", &rmplan.SyncSLORules{
-			App:                         r.instance.Spec.App,
-			Namespace:                   r.instance.TargetNamespace(),
-			Labels:                      r.defaultLabels(),
-			ServiceLevelObjectiveLabels: labels,
-			ServiceLevelObjectives:      slos,
-		}); err != nil {
-			return err
-		}
-	} else {
 		if err := r.applyPlan(ctx, "Delete App SLO Rules", &rmplan.DeleteSLORules{
 			App:       r.instance.Spec.App,
 			Namespace: r.instance.TargetNamespace(),
 		}); err != nil {
 			return err
 		}
-	}
 
 	return nil
 }
